# src/main/resources/swagger/api.yml
openapi: 3.0.1
info:
  title: VAS API
  description: Swagger documentation for the Value Added Services APIs
  version: '1.0'
servers:
  - url: /
security:
  - open_id_scheme:
      - read
      - write
paths:
  /api/dashboard/uploadCsv:
    post:
      tags:
        - Dashboard Controller
      summary: Inserts information from an CSV file into the database, with the associated user
      operationId: uploadFile
      parameters:
        - name: ratingName
          in: header
          required: true
          schema:
            type: string
        - name: year
          in: header
          required: true
          schema:
            type: integer
            format: int32
        - name: type
          in: header
          required: true
          schema:
            type: string
        - name: companyUser
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: CSV file uploaded with success
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '401':
          description: Authentication Required
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '500':
          description: Unexpected Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseMessage'
      security:
        - bearerAuth: []
        - open_id_scheme: []
  /api/dashboard/saveUserRanges:
    post:
      tags:
        - Dashboard Controller
      summary: Saves the current user ranges
      operationId: saveRanges
      parameters:
        - name: companyUserDTO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/RangeDTO'
        required: true
      responses:
        '200':
          description: Ranges saved with success
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '400':
          description: Bad Request
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
  /api/dashboard/saveReports:
    post:
      tags:
        - Dashboard Controller
      summary: Save new Reports that a user can get
      operationId: saveReports
      parameters:
        - name: companyUserDTO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportDTO'
        required: true
      responses:
        '200':
          description: Reports saved with success
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
  /api/dashboard/ratingsByYear:
    get:
      tags:
        - Dashboard Controller
      summary: Retrieves ratings based on inserted year
      operationId: ratingsByYear
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: companyUserDTO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      responses:
        '200':
          description: Ratings of inserted custom year retrieved with success
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataSourceDTO'
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
  /api/dashboard/getWorldMap:
    get:
      tags:
        - Dashboard Controller
      summary: Retrieves a score based on selected ratings, year and current user
      operationId: getDashBoardWorldMap
      parameters:
        - name: ratings[]
          in: query
          required: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/RatingDTO'
        - name: companyUser
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
        - name: year
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        '200':
          description: World map information requested with success
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DashBoardWorldMapDTO'
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
  /api/dashboard/getUserRanges:
    get:
      tags:
        - Dashboard Controller
      summary: Retrieves current user ranges
      operationId: userRanges
      parameters:
        - name: companyUser
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      responses:
        '200':
          description: User Ranges requested with success
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RangeDTO'
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
  /api/dashboard/getTemplate:
    get:
      tags:
        - Dashboard Controller
      summary: Retrieves an CSV template
      operationId: getTemplate
      parameters:
        - name: companyUserDTO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      responses:
        '200':
          description: CSV file retrieved with success
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
                  format: byte
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
  /api/dashboard/getTableInfo:
    get:
      tags:
        - Dashboard Controller
      summary: Retrieves Business partners based on selected ratings, year and current user
      operationId: getAllDashBoardTable
      parameters:
        - name: ratings[]
          in: query
          required: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/RatingDTO'
        - name: year
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: companyUser
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      responses:
        '200':
          description: 'Business partners request with success based on selected variables '
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DashBoardTableDTO'
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
  /api/dashboard/getReportsValueByReport:
    get:
      tags:
        - Dashboard Controller
      summary: Retrieves all Reports that a user can get
      operationId: getReportsValueByReport
      parameters:
        - name: reportDTO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ReportDTO'
        - name: companyUserDTO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      responses:
        '200':
          description: Report values requested with success
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportValuesDTO'
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
  /api/dashboard/getReportsByCompanyUser:
    get:
      tags:
        - Dashboard Controller
      summary: Retrieves all Reports that a user can get
      operationId: getReportsByCompanyUser
      parameters:
        - name: companyUserDTO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      responses:
        '200':
          description: Reports requested with success
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportDTO'
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
  /api/dashboard/getCountryFilterByISO2:
    get:
      tags:
        - Dashboard Controller
      summary: Retrieves all countries in the database filter by ISO CODE 2
      operationId: getCountrys
      parameters:
        - name: companyUserDTO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      responses:
        '200':
          description: Countries requested with success
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryDTO'
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
  /api/dashboard/getCompanyBpns:
    get:
      tags:
        - Dashboard Controller
      summary: Retrieves all Business Partners of a Company
      operationId: getCompanyBpns
      parameters:
        - name: companyUserDTO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      responses:
        '200':
          description: Bpn requested with success
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessPartnerDTO'
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
  /api/dashboard/getBpnCountrys:
    get:
      tags:
        - Dashboard Controller
      summary: Retrieves all countries in the database OF THE Bpns
      operationId: getBpnCountrys
      parameters:
        - name: companyUserDTO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      responses:
        '200':
          description: Countries requested with success
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryDTO'
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
  /api/dashboard/getAllUserBPDMGates:
    get:
      tags:
        - Dashboard Controller
      summary: Retrieves all Gate values that a user can get
      operationId: getAllUserBPDMGates
      parameters:
        - name: companyUserDTO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      responses:
        '200':
          description: Gate values requested with success
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyGatesDTO'
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
...
  /api/dashboard/getAllRatingsForCompany:
    get:
      tags:
        - Dashboard Controller
      summary: Retrieves ratings based on inserted year and Company User
      operationId: getAllRatingsForCompany
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: companyUserDTO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CompanyUserDTO'
      responses:
        '200':
          description: Ratings of inserted custom year retrieved with success
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataSourceDTO'
        '401':
          description: Authentication Required
      security:
        - bearerAuth: []
        - open_id_scheme: []
...
components:
  schemas:
    ResponseMessage:
      type: object
      properties:
        message:
          type: string
    CompanyUserDTO:
      required:
        - companyName
        - email
        - name
      type: object
      properties:
        name:
          type: string
          example: John
        email:
          type: string
          example: John@email.com
        companyName:
          type: string
          example: TestCompany
    RangeDTO:
      required:
        - range
        - value
      type: object
      properties:
        range:
          type: string
          example: Max
          enum:
            - Max
            - Between
            - Min
        value:
          type: integer
          format: int32
          example: 80
        description:
          type: string
          example: Max value
        companyUser:
          $ref: '#/components/schemas/CompanyUserDTO'
...
    ReportValuesDTO:
      type: object
      properties:
        name:
          type: string
          example: Range
        objectValue:
          type: object
...
    RatingDTO:
      type: object
      properties:
        dataSourceName:
          type: string
          example: Fake Rating
        weight:
          type: number
          format: float
          example: 100
...
  securitySchemes:
    open_id_scheme:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://centralidp.dev.demo.catena-x.net/auth/realms/CX-Central/protocol/openid-connect/auth
          tokenUrl: https://centralidp.dev.demo.catena-x.net/auth/realms/CX-Central/protocol/openid-connect/token
          scopes: {}
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT